// backend/src/Interfaces/Providers/IProviders.ts
export interface IProviders {
    id: number;
    name: string;
    tradingName?: string;
    cashforceTax?: string;
    responsibleName?: string;
    responsibleEmail?: string;
    responsiblePosition?: string;
    responsiblePhone?: string;
    responsibleMobile?: string;
    website?: string;
    postalCode?: string;
    address?: string;
    number?: string;
    complement?: string;
    neighborhood?: string;
    city?: string;
    state?: string;
    bank?: string;
    bankAgency?: string;
    account?: string;
    documents?: string;
    phoneNumber?: string;
    situation?: string;
    situationDate?: Date;
    createdAt: Date;
    updatedAt: Date;
    cnpjId?: number;
    email?: string;
  }